Module_6/Advance-Customer-Support-Agent/requirements.txt
python-dotenv>=0.9.9
google-adk>=1.19.0
toolbox-core>=0.5.3
mem0ai>=0.1.118
psycopg2>=2.9.11
tabulate>=0.9.0